Convert server/network/peer.js to ts
Add types.
