+++ server/server.ts
@@ -19,7 +19,7 @@ import * as credentials from "./util/credentials.ts";
 import * as moduleServing from "./modules/serving.js";
 import * as monitor from "./monitoring/monitor.js";
 import * as network from "./network/network.js";
-import * as peer from "./network/peer.js";
+import * as peer from "./network/peer.ts";
 import * as wallGeometry from "./util/wall_geometry.ts";
 import { Control } from "./control.js";
 import { ServerModulePlayer } from "./modules/server_module_player.js";
